Initial commit, add project files.
This code is the most recent worked on. It is from the end of the depth buffering section of the tutorial, but it does not work. It crashes at startup due to invalid read/write errors. The official tutorial code does not crash due to these errors, so code will need to be rolled backed to the beginning of this code section and redone.